geoappfastapi-develop/requirements.txt
fastapi[all]
uvicorn[standard]
psycopg2-binary
geojson
cachetools
python-dotenv
geoalchemy2
SQLAlchemy
python-multipart